Revert "commit e62bc9e55f75ecb41dac3e68cd4b61f9d9501f72"
This reverts commit a16355fbfe4bb1fb3e56be452fbe53619672bc76.

+++ include/net/ip_vs.h
@@ -852,7 +852,7 @@ struct ip_vs_app {
 struct ipvs_master_sync_state {
 	struct list_head	sync_queue;
 	struct ip_vs_sync_buff	*sync_buff;
-	unsigned long		sync_queue_len;
+	int			sync_queue_len;
 	unsigned int		sync_queue_delay;
 	struct task_struct	*master_thread;
 	struct delayed_work	master_wakeup_work;
